chore: split mock objects into different file
