+++ src/lib/stores/drips/utils/__test__/utils.ts
@@ -1,0 +1,3 @@
+import { toWei } from 'web3-utils';
+
+export const eth = (amount: number) => BigInt(toWei(String(amount), 'ether'));
